runs/artifacts/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# Data Processing
pandas==2.2.2
numpy==1.26.4
pyarrow==16.0.0

# HTTP / API clients
httpx==0.27.0
aiohttp==3.9.5
requests==2.31.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0

# Scheduling & Task Queue
apscheduler==3.10.4
celery==5.4.0
redis==5.0.4

# CSV / File handling
python-multipart==0.0.9
openpyxl==3.1.2
chardet==5.2.0

# Validation & Parsing
jsonschema==4.22.0
python-dateutil==2.9.0
pytz==2024.1

# Retry / Resilience
tenacity==8.3.0
circuitbreaker==2.0.0

# Monitoring & Logging
structlog==24.1.0
prometheus-client==0.20.0
rich==13.7.1

# Security
cryptography==42.0.7
python-jose[cryptography]==3.3.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
faker==25.0.0

# CLI
typer==0.12.3
click==8.1.7
